Fix msconvert where tar filename or dataset name has changed
It's easy to change the tar filename after creating it and the dataset name after uploading it. This would break msconvert which assumed the file/folder inside to be named the same as the tar or data set (minus the tar extension). This patch makes it pass msconvert a pathmask based on the datatype. Technically this shortcut (using wildcards) could produce multiple outputs if someone put multiple raw files in a tar. But a more sophisticated approach - which actually tried to pick the real file/folder in the tar - would only improve matters by virtue of throwing a meaningful error if a tar includes multiple raw files.