just chown the complete working dir
in case of the sudo based approach some of the created files
were still owned by root which made planemo fail (bacause
the files could not be removed) and also the final files in
Galaxy's data folder would have the owner root.

So I changed to just chown the complete `working` folder (`./`).