OpenMS place each in/output synlink in unique dir
before each symlink was placed in a dir named by the name of the
corresponding parameter. this still allows for conflicts if several
data sets have the sem identifier.

solution: place symlinks in directory parameter_name/galaxy-dataset-id

usecase: ConsensusID takes the result of different search engines
for the same sample. then the identifiers of the inputs are all the
sample name which leads to a conflict.