make OMS optional parameters with default required in Galaxy
OpenMS uses sets text, int, select, bool parameters that have a default
as optional (required=False), the default value is set implicitly if no
value is given.
This is reasonable for the CLI because one certainly does not want the
user to specify the default manually for all parameters.
For Galaxy tools setting these parameters as required leads to the
equivalent behavior. Assuming required is better because it makes
the implicit setting of parameters more transparent to the user
(in Galaxy the default would be prefilled in the form and at least
one option needs to be selected).