some fixes for searchgui and peptideshaker
1. remove stderr redirection

in case of

```
command 2>> $temp_stderr &&
.... &&
cat $temp_stderr 2>&1;
```

stderr is swallowed if command fails.
this was the case for PS, FC, and IP

in case of

```
command 2>> $temp_stderr &&
.... ;
cat $temp_stderr 2>&1;
```

the last cat is executed even if the command fails.
the exit code of the command line is then the exit
code of the cat, i.e. the tool will always be counted
successful, even if everything fails.

2. fix peptideshaker find exec constructs

Galaxy removes the last ';' from the generated command line.
thus we need to make sure that the last `find ... -exec ... \;`
is not the last command.

3. remove some unnecessay parentheses