fix fasta_cli and ident_param version to searchgui
in the past updates the versions of FC and IP did not change,
which seemed reasonable because the tools did "not change". But actually
the requirement (SG) was updated.

This may result in multiple installations of FC and IP with different
requirements. When installed successively the problem is that
the dependency reolution can only resolve one of the envs so at least
one tool won't have a env and will fail therefore.