use format_source for output
Version 1.0.2 of the tool uses format="input" in its output declaration. This creates an output with data type "input". I figure, the intention is instead to create an output with the same data type as the input. This should fix it, right?

Incidentally, this same error can currently be found in the Tool XML documentation.